--- Resources/en.lproj/Instruments.strings
@@ -1,4 +1,0 @@
-"bass.name" = "Bass";
-"guitar.name" = "Guitar";
-"irish_bouzouki.name" = "Irish Bouzouki";
-"ukulele.name" = "Ukulele";
