--- Resources/es.lproj/Instruments.strings
@@ -1,1 +1,0 @@
-"guitar.name" = "Guitarra";
