--- Resources/es.lproj/Tunings.strings
@@ -1,1 +1,0 @@
-"standard" = "Estándar";
